requirement.txt
asgiref==3.5.2
astroid==2.11.5
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.4
Django @ file:///tmp/build/80754af9/django_1625585912945/work
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
requests==2.28.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.11
wincertstore==0.2
